changing name/location of collectionValidation
